feat(Lotto): 가격을 BigDecimal로 변경
가격을 범위 제약이 없는 BiDecimal로 변경했습니다.

+++ src/main/java/lotto/Lotto.java
@@ -3,6 +3,7 @@ package lotto;
 import static lotto.message.ErrorMessage.LOTTO_IS_NOT_SIX_DIGITS;
 import static lotto.message.ErrorMessage.LOTTO_NUMBER_IS_DUPLICATED;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -10,7 +11,7 @@ import lotto.utils.LottoUtil;
 
 public class Lotto {
     private final List<Integer> numbers;
-    public final static int PRICE = 1000;
+    public final static BigDecimal PRICE = BigDecimal.valueOf(1000);
 
     public Lotto(List<Integer> numbers) {
         validate(numbers);
